organized components into folders and fixed navbar bleeding though viewport when resizing the screen
