+++ src/components/schedule/classView.js
